code 12: Documentation of the catalog of the route of administration of the medication.
