+++ src/main/java/com/project/app_dog_back/application/mapper/IViaAdministracionMapper.java
@@ -1,6 +1,6 @@
 package com.project.app_dog_back.application.mapper;
 
-import com.project.app_dog_back.application.dto.ViaAdministracionDto;
+import com.project.app_dog_back.application.dto.catalog.ViaAdministracionDto;
 import com.project.app_dog_back.domain.model.catalog.ViaAdministracion;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
