+++ src/main/java/com/project/app_dog_back/domain/service/IViaAdministracionService.java
@@ -1,6 +1,6 @@
 package com.project.app_dog_back.domain.service;
 
-import com.project.app_dog_back.application.dto.ViaAdministracionDto;
+import com.project.app_dog_back.application.dto.catalog.ViaAdministracionDto;
 import com.project.app_dog_back.domain.model.catalog.ViaAdministracion;
 
 import java.util.List;
